Changes for proyects domain and services
